backend/requirements.txt
django>=5.2
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
Pillow>=10.0
